Revert "libsodium: update to 1.0.16"
This reverts commit 082668f90f8bc89ba131dd8040badfbba53720ab.

After update of libsodium, dnsmasq-proxy stopped working, have to check why...

+++ release/src-rt-6.x.4708/router/libsodium/src/libsodium/crypto_hash/sha256/cp/hash_sha256_cp.c
@@ -153,10 +153,10 @@ static const uint8_t PAD[64] = { 0x80, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
 static void
 SHA256_Pad(crypto_hash_sha256_state *state, uint32_t tmp32[64 + 8])
 {
-    unsigned int r;
-    unsigned int i;
+    uint64_t r;
+    uint64_t i;
 
-    r = (unsigned int) ((state->count >> 3) & 0x3f);
+    r = (state->count >> 3) & 0x3f;
     if (r < 56) {
         for (i = 0; i < 56 - r; i++) {
             state->buf[r + i] = PAD[i];
